OpacityView: Refactor when it is hidden/shown
When panels are going to be opened the opacityView should be shown. When
panels will be hidden opacityView will be hidden after the animations
are done. This commit also makes it so that contentViewOpacity does it
what it says in that if it is 1 at the end of opening a panel the
contentView will fully show, if it is 0 it will be completely hidden.